NODE-1286-race-condition Refactoring
